Update templates pages
