+++ packages/frontend/app/(application)/settings/templates/add/page.tsx
@@ -1,19 +1,14 @@
 import { Metadata } from 'next';
 import { FC } from 'react';
 
-import { TemplateAddContainer } from '$container/templates/template.add.container';
-import { Layout } from '$layouts/Layout';
+import { TemplateAddContainer } from '$container/templates/TemplateAddContainer';
 
 export const metadata: Metadata = {
   title: 'Add Template',
 };
 
 const AddTemplatePage: FC = () => {
-  return (
-    <Layout title="Add Template">
-      <TemplateAddContainer />
-    </Layout>
-  );
+  return <TemplateAddContainer />;
 };
 
 export default AddTemplatePage;
