+++ packages/frontend/app/(application)/settings/templates/page.tsx
@@ -1,19 +1,14 @@
 import { Metadata } from 'next';
 import { FC } from 'react';
 
-import { TemplateListingContainer } from '$container/templates/template.listing.container';
-import { Layout } from '$layouts/Layout';
+import { TemplateListingContainer } from '$container/templates/TemplateListingContainer';
 
 export const metadata: Metadata = {
   title: 'Templates',
 };
 
 const TemplateListingPage: FC = () => {
-  return (
-    <Layout title="Templates">
-      <TemplateListingContainer />
-    </Layout>
-  );
+  return <TemplateListingContainer />;
 };
 
 export default TemplateListingPage;
